requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-autoslug==1.9.9
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
packaging==25.0
python-telegram-bot==20.7
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.5.0
